// clients/ordersService/models/OrderUpdateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OrderUpdateDto = {
    tenantId?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    companyName?: string | null;
    billingEmail?: string | null;
    addressLine1?: string | null;
    addressLine2?: string | null;
    postalCode?: string | null;
    countryId?: string | null;
    stateId?: string | null;
    cityId?: string | null;
    billingLocationId?: string | null;
    shippingLocationId?: string | null;
    shippingMethodId?: string | null;
    cartId?: string | null;
    userId?: string | null;
    forexRate?: number;
    currencyId?: string | null;
    enrollmentId?: string | null;
    individualId?: string | null;
    organizationId?: string | null;
    totalAmountInUsd?: number;
    totalTaxesInUsd?: number;
    receiverTenantId?: string | null;
    closed?: boolean;
    priceListId?: string | null;
    paymentTermId?: string | null;
    quoteStatus?: string | null;
    effectiveTo?: string | null;
    effectiveFrom?: string | null;
    description?: string | null;
    title?: string | null;
};